requirements-minimal.txt
cryptography
dotenv
fastapi
livekit
livekit-agents
livekit-api
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-openai
livekit-plugins-groq
livekit-plugins-silero
livekit-plugins-sarvam
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-speechify
markupsafe
pydantic
pydantic[email]
pymongo
pyshorteners
requests
sentence-transformers
tqdm
uvicorn
google-genai
python-multipart